Les entités de température sont en nombre entiers #38

+++ custom_components/hilo/sensor.py
@@ -353,7 +353,7 @@ class TemperatureSensor(HiloEntity, SensorEntity):
 
     @property
     def state(self):
-        return str(int(self._device.get_value("current_temperature", 0)))
+        return str(float(self._device.get_value("current_temperature", 0)))
 
     @property
     def icon(self):
